src/test/java/fr/kevin/llps/conf/event/reminder/service/TalkServiceTest.java
package fr.kevin.llps.conf.event.reminder.service;

import fr.kevin.llps.conf.event.reminder.domain.Talk;
import fr.kevin.llps.conf.event.reminder.service.database.MySQLDatabaseService;
import fr.kevin.llps.conf.event.reminder.service.database.mapper.TalkMapper;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;

import java.sql.Connection;
import java.sql.PreparedStatement;
import java.sql.ResultSet;
import java.sql.SQLException;
import java.util.List;

import static fr.kevin.llps.conf.event.reminder.samples.TalkSample.talkList;
import static org.assertj.core.api.Assertions.assertThat;
import static org.mockito.Mockito.*;

@ExtendWith(MockitoExtension.class)
class TalkServiceTest {

    @Mock
    private MySQLDatabaseService databaseService;

    @Mock
    private TalkMapper talkMapper;

    @InjectMocks
    private TalkService talkService;

    @Test
    void shouldGetTalks() throws SQLException {
        List<Talk> expectedTalks = talkList();

        Connection connection = mock(Connection.class);
        PreparedStatement preparedStatement = mock(PreparedStatement.class);
        ResultSet resultSet = mock(ResultSet.class);

        when(databaseService.openConnection()).thenReturn(connection);
        when(connection.prepareStatement(anyString())).thenReturn(preparedStatement);
        when(preparedStatement.executeQuery()).thenReturn(resultSet);
        when(talkMapper.mapToDomainList(resultSet)).thenReturn(expectedTalks);

        List<Talk> talks = talkService.getTalks();

        assertThat(talks).containsExactlyInAnyOrderElementsOf(expectedTalks);

        verifyNoMoreInteractions(databaseService);
    }

}
